server/requirements.txt
fastapi[standard]
uvicorn
psycopg2-binary
pydantic